refactor: 파일 오타 수정(ButLottoForm ➔ BuyLottoForm)
